Upgrade the library to PHPUnit 9
This change upgrades PHPUnit to version 9, so that it can make use of
most of the latest functionality available in that package and not fall
too far behind the curve. The change also upgrades the two other dev
dependencies to the latest versions respectively, for good measure.

I had originally intended to upgrade the package to PHPUnit 10, but that
has a minimum PHP version of 8.1, which might be too far of an upgrade
if this package wants broad PHP support. This change only requires
PHPUnit 7.1, but composer 2 requires a minimum version of PHP 7.2,
which, to me, makes more sense, as both of these version are EOL.